Seperate dry-run of AddData to improve addition of new sim files

+++ Scripts/WorkflowScripts/ProcessInfoFile.py
@@ -10,10 +10,9 @@ Executes standard pipeline of processing for an info file.
    Users of the Databank repository can safely ignore it.
 """
 
-def main(info_file_path): 
+def main(info_file_path:str, dry_run: bool): 
     path_dict = get_databank_paths(NMLDB_ROOT_PATH)
     parent_folder = os.path.dirname(NMLDB_ROOT_PATH)
-
     base_tmp = os.path.join(parent_folder, "databank_workdir")
     work_directory_dry  = os.path.join(base_tmp, "dry")
     work_directory_real = os.path.join(base_tmp, "real")
@@ -23,28 +22,29 @@ def main(info_file_path):
     except OSError as e:
         logger.error(f"Failed to create work directories")
         sys.exit(1)
-
-    run_python_script(
-        path_dict["AddData_path"],
-        args=["-f", info_file_path, "-w", work_directory_dry, "--dry-run"],
-        error_message="AddData dry run failed"
-    )
-    run_python_script(
-        path_dict["AddData_path"],
-        args=["-f", info_file_path, "-w", work_directory_real],
-        error_message="AddData failed"
-    )
-    run_python_script(
-        path_dict["compute_databank_path"],
-        args = ["--nmrpca", "--maicos", "--op", "--thickness","--apl", "--range", "*-0"],
-        error_message="Compute_databank failed"
-    )
-    delete_info_file(info_file_path)
+    if dry_run:
+        run_python_script(
+            path_dict["AddData_path"],
+            args=["-f", info_file_path, "-w", work_directory_dry, "--dry-run"],
+            error_message="AddData dry run failed")
+    else:    
+        run_python_script(
+            path_dict["AddData_path"],
+            args=["-f", info_file_path, "-w", work_directory_real],
+            error_message="AddData failed"
+        )
+        run_python_script(
+            path_dict["compute_databank_path"],
+            args = ["--maicos","*-0"],
+            error_message="Compute_databank failed"
+        )
+        delete_info_file(info_file_path)
 
 #Gets arguments from parser
 def get_args():
     parser = argparse.ArgumentParser(description="Run NMRLipids data pipeline on a YAML info file.")
     parser.add_argument("--info_file_path", required=True, help="Path to the info.yml file")
+    parser.add_argument("--dry-run", action="store_true", help="Only run preanalysis (no compute/commit)")
     return parser.parse_args()
 
 
@@ -54,4 +54,4 @@ if __name__ == "__main__":
     if not info_file_path:
         print("No path provided, exiting.")
         sys.exit(0)
-    main(info_file_path)
+    main(info_file_path,dry_run=args.dry_run)
